create new logged in container
