+++ public/index.styl
@@ -58,8 +58,8 @@ a
 hr
     border 0
     border-top 1px solid $hr-background-color
-    margin-top 20px
     margin-bottom 20px
+    margin-top 20px
 
 code
     background-color $code-background-color
